requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
python-dotenv
youtube-transcript-api